Add PlayerList file
